Refactor Plant View  To Reduce State
This change refactors the components that make up the Plant View Page to reduce state. There are no more callbacks which are passed down in order to pull state back up and instead the lower components emits an action to update the navigation store. This way state is maintained in App and passed down as props.