`getPlantsReadyBetween` does not get all plants ready between two dates
Getting all the plants which are ready by a certain date should account for the harvest time of the plant. There are three cases in which we want to include a calendar item:

1. The ready date falls between the start and end dates

2. The final ready date (the ready date plus the plant's harvest time) falls between the start and end dates

3. The start date falls between the ready date and the final ready date

The final case accounts for the start and end dates being completely contained by the ready date and final ready date interval.