Doctrine does not support robust filtering on date ranges
Doctrine currently does not support, in either the DQL or the query builder, the kind of date querying logic that is needed to properly filter the set of calendar events. Specifically, it does not handle date intervals and is unable to cast between types and so cannot parse an interval string and cast the result to an integer to be later passed to `DATE_ADD`.

Unfortunately, this means that filtering must be done in PHP.